Move shortcuts class to the common library
