+++ trview.app/trview.app.vcxproj
@@ -55,7 +55,6 @@
     <ClCompile Include="Menus\LevelSwitcher.cpp" />
     <ClCompile Include="Menus\MenuDetector.cpp" />
     <ClCompile Include="Menus\RecentFiles.cpp" />
-    <ClCompile Include="Menus\Shortcuts.cpp" />
     <ClCompile Include="Menus\UpdateChecker.cpp" />
     <ClCompile Include="Menus\ViewMenu.cpp" />
     <ClCompile Include="Routing\Route.cpp" />
@@ -129,7 +128,6 @@
     <ClInclude Include="Menus\LevelSwitcher.h" />
     <ClInclude Include="Menus\MenuDetector.h" />
     <ClInclude Include="Menus\RecentFiles.h" />
-    <ClInclude Include="Menus\Shortcuts.h" />
     <ClInclude Include="Menus\UpdateChecker.h" />
     <ClInclude Include="Menus\ViewMenu.h" />
     <ClInclude Include="Routing\Route.h" />
