+++ trview.app/trview.app.vcxproj.filters
@@ -196,9 +196,6 @@
     <ClCompile Include="Lua\Lua.cpp">
       <Filter>Lua</Filter>
     </ClCompile>
-    <ClCompile Include="Menus\Shortcuts.cpp">
-      <Filter>Menus</Filter>
-    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="Camera\Camera.h">
@@ -416,9 +413,6 @@
     </ClInclude>
     <ClInclude Include="Lua\Lua.h">
       <Filter>Lua</Filter>
-    </ClInclude>
-    <ClInclude Include="Menus\Shortcuts.h">
-      <Filter>Menus</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
