+++ trview.common/Windows/Shortcuts.cpp
